Introduce a script to generate a database usable by osv-offline
This script largely builds on top of PR #89 with some necessary
modifications and additions.

The finalized script downloads and extracts the upstream OSV database
and generates additional files for container and RPM vulnerabilities.

Few modifications to the original scripts were necessary:
- Add a retry mechanism for getting the CSAF VEX files. Transient
  networking errors were common during testing and retries shold resolve
  them.
- Add internal ID (_id) to the OSV schema and set it to a random string.
  Every object in the nedb database must have this property.
- Split how the affectedList is generated between RPMs and containers.
- Duplicate affectedList entries for containers to also have registry
  registry.access.redhat.com. This is necessary because the CSAF VEX
  data only tracks images with registry.redhat.io.

This script is expected to be run periodically by a cronjob and
regenerate the DB files onto a persistent volume. Add the new script
to the mintmaker container, so that the cronjob can use the same image.

WARNING: The script now gets the list of advisories from changes.csv.
This is incorrect and should be changed to releases.csv when it becomes
available. The current state is for testing purposes only.