Run migration tool in update executionMode (rework)
KONFLUX-7886

This is a workaround for the story. The pipeline-migration-tool command
line will be short enough with a single Renovate upgrade. However, the
downside is that the migration tool will need to be run as many times as
the total number of dependency updates, which may slow down the whole
Renovate run for a Component repository.