Set replacementName for deprecated Tekton tasks
--- Problem ---

Around June 2024, we started releasing Tekton tasks to the
quay.io/konflux-ci/tekton-catalog namespace.

But we did not stop releasing them to
quay.io/redhat-appstudio-tekton-catalog, because all existing Konflux
users at the time would have the redhat-appstudio-tekton-catalog
references in their .tekton/ yaml files.

We did not know of any good way to migrate the users over to the new
konflux-ci/tekton-catalog references, so we just kept releasing to both
places.

Now, with the decentralization of build-definitions in progress, we are
starting to release Tekton tasks via Konflux, which will make it hard
(if not impossible) to keep releasing to the old location.

--- Solution ---

Configure replacementName [1] for all tasks that exist both in
redhat-appstudio-tekton-catalog and in konflux-ci/tekton-catalog. This
will cause Mintmaker to send out PRs that automatically migrate users
away from redhat-appstudio-tekton-catalog.

Note: I'm aware that replacementNameTemplate [2] exists, and that it
would make this commit roughly 6 lines long. Unfortunately, Renovate
seems to have a bug that makes replacementNameTemplate unusable in this
case: https://github.com/renovatebot/renovate/discussions/37744

[1]: https://docs.renovatebot.com/configuration-options/#replacementname
[2]: https://docs.renovatebot.com/configuration-options/#replacementnametemplate

Signed-off-by: Adam Cmiel <acmiel@redhat.com>