Fix job launch metrics
This commit fixes the job launch metrics, which became inoperative after
changes to the pipelinerun controller.

Previously, the metrics on job launching/scheduling were called during the
pipelinerun controller reconcile loop. However, since Keue was adopted, this
controller loop was removed, since it does not need to manage the pipeline
run launches anymore.

This commit reintroduces the metric in the DUC controller, and makes the
metric opperative once again.